apps/content/pyproject.toml
[project]
name = "content"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Willem Meints", email = "1550763+wmeints@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dapr-ext-grpc>=1.15.0",
    "dapr>=1.15.0",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.11.4",
    "pymongo>=4.13.0",
    "strawberry-graphql>=0.268.1",
    "typer>=0.15.4",
    "uvicorn>=0.34.2",
]

[project.scripts]
agenttalks-content="agenttalks.content.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agenttalks"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.10",
]
